+++ src/recipes/like/like.service.ts
@@ -1,34 +1,38 @@
 import { Injectable } from '@nestjs/common'
-import { CreateRecipeLikeInput } from 'src/recipes/like/inputs/create-recipe-like.input'
 import { PrismaService } from 'src/prisma/prisma.service'
 
 @Injectable()
 export class LikeService {
 	constructor(private readonly prisma: PrismaService) {}
-	create(userId: string, data: CreateRecipeLikeInput) {
-		return this.prisma.recipeLike.create({
-			data: {
-				userId: userId,
-				recipeId: data.recipeId
+	async toggle(userId: string, recipeId: string) {
+		const data = {
+			recipeId,
+			userId
+		}
+
+		const like = await this.prisma.recipeLike.findUnique({
+			where: {
+				recipeId_userId: data
 			}
 		})
+
+		if (like) {
+			return this.prisma.recipeLike.delete({
+				where: {
+					recipeId_userId: data
+				}
+			})
+		} else {
+			return this.prisma.recipeLike.create({
+				data
+			})
+		}
 	}
 
-	findCount(data: CreateRecipeLikeInput) {
+	findCount(recipeId: string) {
 		return this.prisma.recipeLike.count({
 			where: {
-				recipeId: data.recipeId
-			}
-		})
-	}
-
-	remove(userId: string, data: CreateRecipeLikeInput) {
-		return this.prisma.recipeLike.delete({
-			where: {
-				recipeId_userId: {
-					recipeId: data.recipeId,
-					userId: userId
-				}
+				recipeId
 			}
 		})
 	}
